pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.15.0",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.9.0",
    "npm>=0.1.1",
    "npx>=0.1.6",
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "pywinauto>=0.6.9",
]
